+++ CHANGELOG.md
@@ -4,6 +4,7 @@ ENHANCEMENTS:
 BUG FIXES:
 * fix typo in errors and commits messages
 * [workflows] fix compile freebsd/arm64 on release
+* fix rule/policy with space in name for application-services in resource junos_security_policy
 
 ## 1.2.1
 ENHANCEMENTS: